backend/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
pandas>=2.2
openpyxl>=3.1
xlrd>=2.0
